+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Release Versions:
 
 ## Upcoming changes (in development)
 
+- Remove rows_ and cols_ and improve constructors (#100)
 - Revise clproto to remove StateType and timestamp from State message (#104)
 - Raise exception in getters if state is empty (#97)
 - Bind error objects of all library modules with Pybind (#98)